Don't describe-param when doing quiet read
The describe-parameter feature requires access to the root path of
the store.  When using multiple services, we have a use case where
we only provide the necessary paths to each ecs-task role that the
service needs.

This works fine for exec calls, which use parameters-by-path, but
when we try to read specific keys, this fails because it tries
to describe all parameters, which the ecs-task-role does not have
permission to do.

This feature makes the 'quiet' flag on the read operation truly
quiet - as in it doesn't bother to go get the metadata that is then
thrown away.  This also allows you to trim the permissions down
for a specific role, which works well for services that don't
care about the key metadata.